Refactor localEventBase resource arrival
