+++ _common/gameModel/libs/ClientScript/game/common/impacts/implementation/activationImpact.ts
@@ -1,10 +1,7 @@
 import { ITemplateDescriptor } from '../../actions/actionTemplateDescriptor/templateDescriptor';
 import { ChoiceDescriptor } from '../../actions/choiceDescriptor/choiceDescriptor';
 import { Uid } from '../../interfaces';
-import {
-  LocalEventBase,
-  SourceType,
-} from '../../localEvents/localEventBase';
+import { LocalEventBase, SourceType } from '../../localEvents/localEventBase';
 import { BuildStatus, MapEntityDescriptor } from '../../mapEntities/mapEntityDescriptor';
 import { MainSimulationState } from '../../simulationState/mainSimulationState';
 import { Trigger } from '../../triggers/trigger';
