+++ _common/gameModel/libs/ClientScript/game/common/impacts/implementation/choiceEffectSelectionImpact.ts
@@ -1,8 +1,5 @@
 import { Uid } from '../../interfaces';
-import {
-  LocalEventBase,
-  SourceType,
-} from '../../localEvents/localEventBase';
+import { LocalEventBase, SourceType } from '../../localEvents/localEventBase';
 import { MainSimulationState } from '../../simulationState/mainSimulationState';
 import { ImpactBase } from '../impact';
 import { SelectChoiceEffectLocalEvent } from '../../localEvents/localEventBaseActivable';
