+++ _common/gameModel/libs/ClientScript/game/common/localEvents/localEventBaseActivable.ts
