+++ _common/gameModel/libs/ClientScript/game/common/tasks/taskBaseEvacuation.ts
@@ -12,10 +12,7 @@ import {
   TranslationKey,
 } from '../baseTypes';
 import { EvacuationSquadDefinition } from '../evacuation/evacuationSquadDef';
-import {
-  MovePatientLocalEvent,
-
-} from '../localEvents/localEventBase';
+import { MovePatientLocalEvent } from '../localEvents/localEventBase';
 import { getLocalEventManager } from '../localEvents/localEventManager';
 import { RadioType } from '../radio/communicationType';
 import * as RadioLogic from '../radio/radioLogic';
@@ -26,7 +23,8 @@ import { TaskBase, TaskType } from './taskBase';
 import { AddRadioMessageLocalEvent } from '../localEvents/localEventBaseRadio';
 import {
   AssignResourcesToWaitingTaskLocalEvent,
-  MoveResourcesAtArrivalLocationLocalEvent, MoveResourcesLocalEvent,
+  MoveResourcesAtArrivalLocationLocalEvent,
+  MoveResourcesLocalEvent,
 } from '../localEvents/localEventBaseResources';
 
 // -------------------------------------------------------------------------------------------------
