+++ _common/gameModel/libs/ClientScript/game/common/triggers/trigger.ts
@@ -3,13 +3,14 @@ import { triggerLogger } from '../../../tools/logger';
 import { getTriggers } from '../../loaders/triggerLoader';
 import { convertToLocalEvents, Impact } from '../impacts/impact';
 import { IActivableDescriptor, IDescriptor, Indexed, Typed } from '../interfaces';
-import {
-  LocalEventBase,
-} from '../localEvents/localEventBase';
+import { LocalEventBase } from '../localEvents/localEventBase';
 import { getTriggerActivable, TriggerActivable } from '../simulationState/activableState';
 import { MainSimulationState } from '../simulationState/mainSimulationState';
 import { Condition, evaluateCondition } from './condition';
-import { ChangeActivableStatusLocalEvent, IncrementCountLocalEvent } from '../localEvents/localEventBaseActivable';
+import {
+  ChangeActivableStatusLocalEvent,
+  IncrementCountLocalEvent,
+} from '../localEvents/localEventBaseActivable';
 
 /**
  * A trigger is a collection of conditions and impacts.
